rename plotting_utils function to "set_pyplot_defaults"
